+++ src/Plugin.cs
@@ -3,6 +3,7 @@ using System.Reflection;
 using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
+using BetterCartographyTable.Model;
 using HarmonyLib;
 using Jotunn.Utils;
 using UnityEngine;
@@ -19,10 +20,12 @@ public class Plugin : BaseUnityPlugin
   private const string ModName = "BetterCartographyTable";
   private const string ModVersion = "0.4.3";
 
+  private static ConfigEntry<SharingMode> s_mapExplorationSharingMode;
   private static ConfigEntry<KeyCode> s_modifierKey;
   private static ConfigEntry<Color> s_publicPinsColor;
 
   public static new ManualLogSource Logger;
+  public static SharingMode MapExplorationSharingMode => s_mapExplorationSharingMode.Value;
   public static KeyCode ModifierKey => s_modifierKey.Value;
   public static bool IsModifierKeyPressed => Input.GetKey(ModifierKey);
   public static Color PublicPinsColor => s_publicPinsColor.Value;
@@ -30,6 +33,11 @@ public class Plugin : BaseUnityPlugin
   public void Awake()
   {
     Logger = base.Logger;
+    s_mapExplorationSharingMode = Config.Bind("Behaviour", "Map exploration sharing mode", SharingMode.Public, @"Whether to share your own map exploration to cartography tables you interact with.
+No matter your choice, you will always receive other players' map exploration data (if anything was shared to the table).
+  - Private: never share your map exploration
+  - Public: always share your map exploration with all cartography tables (both public and guild)
+  - Guild: only share your map exploration with guild tables");
     s_modifierKey = Config.Bind("Keys", "Modifier key", KeyCode.LeftShift, "Modifier key to use for interacting with public or guild pins on the cartography table.");
     s_publicPinsColor = Config.Bind("UI", "Public pins color", Color.green, "Color to use for public pins.");
     SetUpConfigWatcher();