+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add config toggle for map exploration sharing.
+
 ## [0.4.3] - 2024-06-01
 
 ### Changed